68696 refactor map classes to avoid conflicts
